[ecore 016] Moved class MessageUtil do package com.ecore.roles.utils
